server/requirements.txt
discord_py_interactions==5.13.2
